+++ src/components/wt-app-header/wt-app-header.vue
@@ -7,7 +7,7 @@
 <script setup></script>
 
 <style lang="scss">
-@import './variables.scss';
+@use './variables.scss';
 </style>
 
 <style lang="scss" scoped>