+++ src/components/wt-app-header/wt-header-actions.vue
@@ -131,7 +131,7 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-@import '../../../src/css/main.scss';
+@use '../../css/main.scss';
 
 .wt-header-actions {
   position: relative;