+++ src/components/wt-avatar/wt-avatar.vue
@@ -121,7 +121,7 @@ const badgeColorVar = computed(() => {
 </script>
 
 <style lang="scss">
-@import './variables.scss';
+@use './variables.scss';
 </style>
 
 <style lang="scss" scoped>