+++ src/components/wt-context-menu/wt-context-menu.vue
@@ -82,11 +82,11 @@ function handleOptionClick({ option, index, hide }) {
 </script>
 
 <style lang="scss">
-@import './variables.scss';
+@use './variables.scss';
 </style>
 
 <style lang="scss" scoped>
-@import '../../../src/css/main.scss';
+@use '../../css/main.scss';
 
 .wt-context-menu {
   line-height: 0;