+++ src/components/wt-divider/wt-divider.vue
@@ -18,7 +18,7 @@ export default {
 </script>
 
 <style lang="scss">
-@import './variables.scss';
+@use './variables.scss';
 </style>
 
 <style lang="scss" scoped>