+++ src/components/wt-navigation-bar/wt-navigation-bar.vue
@@ -209,11 +209,11 @@ export default {
 </script>
 
 <style lang="scss">
-@import './variables.scss';
+@use './variables.scss';
 </style>
 
 <style lang="scss" scoped>
-@import '../../../src/css/main.scss';
+@use '../../css/main.scss';
 
 .wt-navigation-bar__menu-btn {
   display: block;