+++ src/components/wt-timepicker/wt-timepicker.vue
@@ -182,7 +182,7 @@ export default {
 </script>
 
 <style lang="scss">
-@import './variables.scss';
+@use './variables.scss';
 </style>
 
 <style lang="scss" scoped>